requirements.txt
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
newspaper3k==0.2.8
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.32.0
sseclient-py==1.8.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4
torch==2.2.0
scikit-learn==1.4.0
python-telegram-bot==20.7
google-generativeai==0.3.2
python-multipart==0.0.9
lxml-html-clean==0.4.2
setuptools
httpx==0.27.0
